requirements.txt
openai==0.27.4
requests==2.31.0
python-dotenv==1.0.0
requests_oauthlib==1.3.1
pymongo==4.4.1
instagrapi==2.0.0
Pillow==10.1.0
google-cloud-storage==2.12.0